requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
gradio
pandas
scikit-learn
rank_bm25

